Re-enable saving Fcalls in a chan and fix a race
I had not realized the Go runtime was still so bad at allocation, so had stopped caching Fcalls.
But when Msize is 1M, the throughput goes to pieces, with fcall allocation taking 1/4 millisecond
on average (!).

I initially disabled it because of a race between send() and receive(), which is now fixed.
I don't like the fix but I think the whole send/recv thing needs a rework. This will do for now
and gets our performance back with 1M msize connections.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>