Fix encoding to so we can stop counting bytes.
If we're going to add 9p2000.L we need to stop counting bytes.

This change lets us do that. We can now just call an Encode function
with the list of elements of a packet and it will create a correctly-encoded
9P packet.

We also seperate the Twrite packet and data for better performance.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>