Test that a server connection has received a Tversion before anything else
We were improperly not checking for a Tversion as the first operation one must do
before any others; as a result we had a bad server and lots of bad tests.

We now require one Tversion before anything else. We add a test for this,
by calling Tattach on a non-Tversioned connection. That used to pass, now it
correctly fails.

All other tests are changed to use Mount now.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>