Minor code fix: have bench print date/args; nullfs now creates read data for null
I had hoped to depend on ref for the benchmarking but it's useful to
- have it print the date
- have it print the benchmark type
- have it print args so you can do, e.g., ./bench before zero copy read
and have those args appear in the file as a comment.

For nullfs, it's more realistic to allocate an array and set the Data to point
to it; this stands in as the overhead of a real read.

We include a benchmark run for pre zero-copy read.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>